pyproject.toml
[project]
name = "roommate"
version = "0.1.0"
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-homeassistant-custom-component",
    "pytest-cov",
    "ruff",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "LOG",    # flake8-logging
    "RUF",    # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["custom_components.roommate"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
